% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode an Address}
\usage{
geocode(address.string)
}
\arguments{
\item{address.string}{character vector of addresses to geocode.}
}
\value{
A data frame with probable address locations.
}
\description{
Get Lat/Lon Cooridinates for an Address.
}
\details{
This function uses the \code{findAddressCandidates} utility at
\href{http://geocode.arcgis.com/arcgis/rest/services/World/GeocodeServer}{Arcgis.com}
to get coordinates for a single address.
}
\examples{
address <- "1600 Pennsylvania Ave NW, Washington, DC 20500"
\dontrun{
g <- geocode(address)
point.matrix <- matrix(
  c(
    as.numeric(g$location[1,1]),
    as.numeric(g$location[1,2])
  ),
  nrow=1
)
point.df <- data.frame(
    pt.type="White House"
)
mymap <- ol_map(
   center=c(-77.03196,38.89037),
   zoom=12
) +
    streetmap() +
ol_geom_point(
    point.matrix,
    name="Points of Interest",
    marker="pin",
    toggle.control=TRUE,
    tooltip=point.df$pt.type
)

# Output to file and view
ol_map2HTML(mymap,'map.html')
browseURL('map.html')
}
}
\seealso{
\code{\link{ol_geom_point}},
\code{\link{ol_geom_icon}},
}
