% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ol_map_base.R
\name{nga_basemap}
\alias{nga_basemap}
\title{NGA Basemap Layer}
\usage{
nga_basemap(basemap.identifier = "WSM", name = NULL,
  toggle.control = FALSE)
}
\arguments{
\item{basemap.identifier}{character indicating which NGA mapserver to use.
See 'Available Base Maps'.}

\item{name}{character layer name.}

\item{toggle.control}{logical.  If \code{TRUE}, a checkbox will appear on the
map allowing the viewer to toggle its visibility in the browser.}
}
\value{
A \code{Layer.ArcGIS} S3 object.
}
\description{
Create a basemap layer linking to an NGA ArcGIS mapserver.
}
\details{
Creates and returns an OpenLayers ArcGIS Tile layer that sources a
map server hosted at \url{https://home.gvs.nga.mil}.  These map servers
are owned by the US Government and require authentication.  If the 
\code{basemap.identifier} parameter is unrecognized the function will
default to the NGA OpenStreetMap map server.
}
\section{Available Base Maps}{

The following basemap.identifiers are currently supported by this method.
\tabular{ll}{
"ABM" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/CanvasMaps/Analytic_Basemap/MapServer}{Analytic Base Map}\cr
"LightGray" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/CanvasMaps/LightGray/MapServer}{Analytic Base Map (Light Gray)}\cr
"Light_LightGray" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/CanvasMaps/Lite_LightGray/MapServer}{Analytic Base Map (Light Light Gray)}\cr
"LightMidnight" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/CanvasMaps/Lite_Midnight/MapServer}{Analytic Base Map (Light Midnight)}\cr
"Light_Slate" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/CanvasMaps/Lite_Slate/MapServer}{Analytic Base Map (Light Slate)}\cr
"Midnight" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/CanvasMaps/Midnight/MapServer}{Analytic Base Map (Midnight)}\cr
"Slate" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/CanvasMaps/Slate/MapServer}{Analytic Base Map (Slate)}\cr
"CARDG" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/Basemap/ScannedMaps/MapServer}{Scanned CARDG Maps}\cr
"DNC" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/Basemap/DNC/MapServer}{Digital Nautical Charts}\cr
"Imagery" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/Basemap/NGA_World_Imagery_2D/MapServer}{Satellite Imagery}\cr
"Hillshade" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/Basemap/NGA_Hillshade_2D/MapServer}{Hillshade Map}\cr
"ShadedRelief" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/Basemap/NGA_ShadedRelief_2D/MapServer}{Shaded Relief Map}\cr
"TintedHillshade" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/Basemap/NGA_Tinted_Hillshade/MapServer}{Tinted Hillshade Map}\cr
"WorldBoundaries" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/Basemap/World_Boundaries_2D/MapServer}{World Boundaries (WSM)}\cr
"WorldBoundaries_Places" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/Basemap/World_Boundaries_Places_2D/MapServer}{World Boundaries, Places (WSM)}\cr
"WorldPlaceNames" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/Basemap/World_Place_Names_2D/MapServer}{World Place Names (WSM)}\cr
"WorldTransportation" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/Basemap/World_Transportation_2D/MapServer}{World Transportation (WSM)}\cr
"WorldCities" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/SampleWorldCities/MapServer}{Sample World Cities}\cr
"WSM" \tab \href{https://maps.gvs.nga.mil/arcgis/rest/services/Basemap/World_StreetMap_2D/MapServer}{World Street Map}\cr
}
}

\examples{
mymap <- ol_map(
    nga.olsource = TRUE ## Not required; can also use public OpenLayers
                        ## javascript source with NGA mapservers.
)
base.layer <- nga_basemap('Midnight')
mymap <- mymap + base.layer
## Not run
# ol_map2HTML(mymap,"SanDiegoMidnight.html")
# browseURL("SanDiegoMidnight.html")
}
\seealso{
\code{\link{ol_map}}, 
\code{\link{+.Ol.Map}}, 
\code{\link{public_arcgis_basemap}}, 
\code{\link{public_OSM_basemap}}, 
\code{\link{user_arcgis_basemap}}
}
