% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ol_map_base.R
\name{ol_aes}
\alias{ol_aes}
\title{Aesthetic Mappings}
\usage{
ol_aes(...)
}
\arguments{
\item{...}{comma-separated mappings of the form 'aesthetic=variable'.  Available aesthetics for mapping
are layer specific and are listed in the documentation for each layer type.  Unavailable or 
unrecognized aesthetics are ignored.  Variables must correspond to names in the layer's input data.frame,
otherwise an error is thrown.}
}
\value{
A list of aesthetic mappings.
}
\description{
Map variables to layer aesthetics.
}
\details{
This function replicates a subset of the functionality of the ggplot2 \code{aes} function.
It \emph{does not} allow for variable transformations or functions of multpile varuables.
These operations must be completed a priori by the user.
}
\examples{
polygon.matrix1 <- matrix(
    c(
        -80.385+c(0,0.05,0.05,0,0),
        25.782618+c(0,0,0.05,0.05,0)
    ),
    ncol=2
)
polygon.matrix2 <- matrix(
    c(
        -80.34+c(0,0.05,0.025,0),
        25.73++c(0,0,0.025*sqrt(3),0)
    ),
    ncol=2
)
polygon.list<-list(polygon.matrix1,polygon.matrix2)
polygon.df <- data.frame(shape=c("rectangle","triangle"),no=c(1,2))
miami.OSM.basemap <- ol_map(
    center=c(-80.385790,25.782618),
    zoom=9,
    map.heading="Miami Shapes",
    map.note="Note: Mouseover popup values are 
        independent of shape size &amp; color."
    ) + 
   nga_basemap('WSM') 
polygon.layer <- ol_geom_polygon(
    polygon.list,
    mapping=ol_aes(
        fill=no,
        lwd=shape
    ),
    df=polygon.df,
    name="Miami Polygons",
    toggle.control=TRUE,
    tooltip=polygon.df$no
)
polygon.fill.scale <- ol_scale_fill_discrete(
    c("1"="red","2"="green"),
    opacity=0.5,
    display=TRUE,
    name="Number"
)
polygon.linewidth.scale <- ol_scale_lwd_discrete(
    display=TRUE,
    name="Shape"
)
polygons.over.miami <- miami.OSM.basemap + 
    polygon.layer + 
    polygon.fill.scale +
    polygon.linewidth.scale

## Not Run
# ol_map2HTML(polygons.over.miami,'miami_polygons.html')
# browseURL("miami_polygons.html")
}
\seealso{
\code{\link{ol_geom_polygon}}, 
\code{\link{ol_geom_line}}, 
\code{\link{ol_geom_point}}, 
\code{\link{ol_geom_icon}}, 
\code{\link{ol_geom_circle}}
}
