% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_circle.R
\name{ol_geom_circle}
\alias{ol_geom_circle}
\title{OpenLayers Circle Layer}
\usage{
ol_geom_circle(circle.obj, mapping = ol_aes(), name = NULL,
  df = NULL, toggle.control = FALSE, fill = "#00FF0090",
  fill.opacity = 0.5, lwd = 1, ol.lty = list(), color = "#000000",
  label = NULL, label.params = list(), tooltip = NULL,
  tooltip.params = list())
}
\arguments{
\item{circle.obj}{matrix containing three columns: center longitude,
center latitude, and radius respectively.  Each row yields a
single \code{circle} feature in the resulting layer.}

\item{mapping}{list created by ol_aes.}

\item{name}{character Layer name.}

\item{df}{data.frame with same number of rows as \code{circle.obj}.
Used for aestheic mapping.}

\item{toggle.control}{logical indicating whether this layer will have
a visibility toggle.}

\item{fill}{character color string, or vector of color strings.  Used
only if no \code{fill} aesthetic is provided in \code{mapping}}

\item{fill.opacity}{numeric in [0,1]. Controls circle opacity if
no opacity provided in \code{fill} or \code{fill} aesthetic.}

\item{lwd}{numeric circle border width. Used only if no \code{lwd}
aesthetic is provided in \code{mapping}}

\item{ol.lty}{(experimental) numeric vector with length > 1, or
\code{list} of such vectors. Used only if no \code{ol.lty} aesthetic is
provided in \code{mapping}.  See OpenLayers
\href{https://openlayers.org/en/latest/apidoc/module-ol_style_Stroke-Stroke.html}{ol/style/Stroke Documentation},
'lineDash' property for more information.}

\item{color}{character border color string, or vector of color strings.  Used
only if no \code{color} aesthetic is provided in \code{mapping}}

\item{label}{character vector of length \code{nrow(circle.obj)} of
feature labels.}

\item{label.params, tooltip.params}{named lists (e.g., \code{list(property=value)}) of
label and tooltip position and format parameters.  See \link{ol_geom_polygon} documentation.}

\item{tooltip}{character vector of length \code{nrow(circle.obj)} of
feature tooltip popups.}
}
\value{
A list object of class \code{Layer.Circle}.
}
\description{
Function to create a circle layer to add to an OpenLayers Map object.
}
\details{
This function stores the data required to generate an OpenLayers
vector layer with features using \code{circle}
geometries.
See OpenLayers \href{https://openlayers.org/en/latest/apidoc/module-ol_geom_Circle-Circle.html}{Circle Documentation} for details.
}
\section{Aesthetics}{

\itemize{
\item \code{fill}
\item \code{color}
\item \code{lwd}
\item \code{ol_lty} (experimental; See OpenLayers
\href{https://openlayers.org/en/latest/apidoc/module-ol_style_Stroke-Stroke.html}{ol/style/Stroke Documentation},
'lineDash' property for more information.)
}
}

\examples{
miami.circles <- matrix(
    c(
        -80.885+runif(10), #Longitudes
        25.282618+runif(10), #Latitudes
        rnorm(10,2000,500) # Radii in meters
    ),
    ncol=3
)
aesthetic.df <- data.frame(
    type=sample(c("A","B"),10,replace=TRUE),
    value=runif(10)*10
)
miami.OSM.basemap <- ol_map(
    center=c(-80.385790,25.782618),
    zoom=9
    ) +
   streetmap()
circle.layer<-ol_geom_circle(
        miami.circles,
        df = aesthetic.df,
        mapping=ol_aes(fill=type),
        lwd=2,
        name="Meaningless Miami Circles",
        toggle.control=TRUE,
        color="#000000FF",
        tooltip=sprintf("\%1.2f",aesthetic.df$value)
        )
circle.fill <- ol_scale_fill_discrete(
        display=TRUE,
        preserve.opacity=TRUE
    )
circles <- miami.OSM.basemap + circle.layer + circle.fill

\dontrun{
# Output to file and view
ol_map2HTML(
  circles,
  'miami_circles.html',
  map.heading="Miami Shapes",
  map.note="Note: Mouseover popup values are
    independent of shape size &amp; color."
)
browseURL("miami_circles.html")
}
}
\seealso{
\code{\link{ol_aes}},
\code{\link{ol_map}},
\code{\link{ol_geom_polygon}}
}
