% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_heatmap.R
\name{ol_geom_heatmap}
\alias{ol_geom_heatmap}
\title{OpenLayers Heatmap Layer}
\usage{
ol_geom_heatmap(point.obj, name = NULL, toggle.control = FALSE,
  gradient = NULL, opacity = 1, radius = 8, blur = 15,
  shadow = 250, weight.values = NULL)
}
\arguments{
\item{point.obj}{SpatialPointsDataframe, SpatialPoints, or a matrix 
containing columns of point longitudes and latitudes, respectively.}

\item{name}{character Layer name.}

\item{toggle.control}{logical indicating whether this layer will have
a visibility toggle.}

\item{gradient}{character color gradient of heatmap.  See OpenLayers 
\href{http://geoadmin.github.io/ol3/apidoc/ol.layer.Heatmap.html}{Heatmap Documentation}
Enclose gradient array in single character string.}

\item{opacity}{numeric Heatmap opacity.  See OpenLayers 
\href{http://geoadmin.github.io/ol3/apidoc/ol.layer.Heatmap.html}{Heatmap Documentation}.}

\item{radius}{numeric Heatmap radius size in pixels.See OpenLayers 
\href{http://geoadmin.github.io/ol3/apidoc/ol.layer.Heatmap.html}{Heatmap Documentation}.}

\item{blur}{numeric Heatmap blur.  See OpenLayers 
\href{http://geoadmin.github.io/ol3/apidoc/ol.layer.Heatmap.html}{Heatmap Documentation}.}

\item{shadow}{numeric Heatmap shadow.  See OpenLayers 
\href{http://geoadmin.github.io/ol3/apidoc/ol.layer.Heatmap.html}{Heatmap Documentation}.}

\item{weight.values}{numeric vector of weights to be assigned to the
points in \code{point.obj}.  Values should be in [0,1].}
}
\value{
A list object of class \code{Layer.HeatMap}.
}
\description{
Function to create a Heatmap layer to add to an OpenLayers Map object.
}
\details{
This function stores the data required to generate an OpenLayers 
vector layer with features using \code{Point} 
geometries.  
See OpenLayers \href{http://geoadmin.github.io/ol3/apidoc/ol.layer.Heatmap.html}{Heatmap Documentation}
for details.
}
\section{Aesthetics}{

\itemize{
\item \code{fill}
\item \code{size}
}
}

\examples{
heatmap.pts <- matrix(
    c(
        rnorm(100,-80.385,1), #Miami Longitudes
        rnorm(100,-117.1611,3), #San Diego Longitudes
        rnorm(100,25.782618,1), #Miami Latitudes
        rnorm(100,32.7157,3) # San Diego Latitudes
    ),ncol=2
)
mymap <- ol_map(
    center=c(-98.5,28.5),
    zoom=4
) + 
    streetmap() +
    ol_geom_heatmap(
        heatmap.pts,
        name="Random Heatmap",
        toggle.control=TRUE,
        opacity=0.25
        )
\dontrun{
# Write to file and view in browser
ol_map2HTML(
  mymap,
  "heatmap.html",
  map.note="Heatmap of random points centered on Miami and San Diego."
)
browseURL("heatmap.html")
}
}
\seealso{
\code{\link{ol_map}}, 
\code{\link{ol_geom_polygon}}, 
\code{\link{ol_geom_circle}},
\code{\link{ol_geom_point}},
\code{\link{ol_geom_icon}}
}
