% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_icon.R
\name{ol_geom_icon}
\alias{ol_geom_icon}
\title{OpenLayers Icon Layer}
\usage{
ol_geom_icon(point.obj, src.img = NULL, mapping = ol_aes(),
  name = NULL, df = NULL, toggle.control = FALSE,
  icon.size.scalar = "autoscale", src.img.width = NULL,
  target.icon.width = 30, size.scale.lims = c(0.5, 1.25),
  label = NULL, label.params = list(), tooltip = NULL,
  tooltip.params = list(), deployment.image.path = NULL)
}
\arguments{
\item{point.obj}{SpatialPointsDataframe, SpatialPoints, or a matrix 
containing columns of point longitudes and latitudes, respectively.}

\item{src.img}{character vector of image file paths.}

\item{mapping}{list created by ol_aes. Used for aestheic mapping.}

\item{name}{character Layer name.}

\item{df}{data.frame with same number of rows as \code{point.obj} coordinate
matrix.}

\item{toggle.control}{logical indicating whether this layer will have
a visibility toggle.}

\item{icon.size.scalar}{numeric scalar vector or 'autoscale'.
The width of the icon on the map will be scaled by this input from the original
image width.  The default 'autoscale' uses the png, jpeg, or tiff package to
scale each image to \code{target.icon.width}.}

\item{src.img.width}{numeric vector of widths of user-supplied images, in pixels.
If \code{icon.size.scalar} is not supplied and \code{src.img.width} is not provided,
image widths will be detected using the png, jpeg, or tiff package.}

\item{target.icon.width}{numeric desired width of icons on map, in pixels.  Only used
if \code{icon.size.scalar} is 'autoscale'.}

\item{size.scale.lims}{numeric vector containing the minimum and maximum image scaling
for size aesthetic mappings.  A value of 1 results renders the image at the default size,
determined by \code{target.icon.width} or \code{icon.size.scalar}.}

\item{label}{character vector of point feature labels.}

\item{label.params, tooltip.params}{named lists (e.g., \code{list(property=value)}) of 
label and tooltip position and format parameters.  See \link{ol_geom_polygon} documentation.}

\item{tooltip}{character vector of point feature tooltip popups.}

\item{deployment.image.path}{character giving the path where the icon images will be located on 
the server, if different from the path in \code{img.src}.}
}
\value{
A list object of class \code{Layer.SpatialIcon}.
}
\description{
Function to create a point-icon layer to add to an OpenLayers Map object.
}
\details{
This function stores the data required to generate an OpenLayers 
vector layer with features using \code{Point} 
geometries and user-supplied point icons.
}
\section{Aesthetics}{

\itemize{
\item \code{icon}
\item \code{size}
}
}

\examples{
some.r.servers <- matrix(
    c(
        144.964,-37.798,
        -122.920,49.278,
        121.494,31.307,
        25.083,35.307,
        -21.930,64.149,
        11.877,45.407,
        -99.200,19.345,
        5.322,60.388,
        -8.224,39.400,
        -8.616,41.147,
        -73.953,40.768,
        20.304,63.821,
        8.548,47.376,
        33.031,35.247,
        -78.938,36.001,
        -123.279,44.564,
        -96.797,32.777
    ),
    byrow=TRUE,
    ncol=2
)
r.server.names <- c(
    'School of Mathematics and Statistics, University of Melbourne',
    'Simon Fraser University, Burnaby',
    'Shanghai University',
    'University of Crete',
    'Marine Research Institute',
    'University of Padua',
    'Instituto Tecnologico Autonomo de Mexico',
    'University of Bergen',
    'RadicalDevelop, Lda',
    'University of Porto',
    'Four Dots',
    'Academic Computer Club, Umeå University',
    'ETH Zurich',
    'Middle East Technical University Northern Cyprus Campus, Mersin',
    'Duke University, Durham, NC',
    'Oregon State University',
    'Revolution Analytics, Dallas, TX'
)
r.icon <- "https://www.r-project.org/Rlogo.png"
## If width is not provided image must be local
## and png package must be installed.
r.icon.width <- 200 
r.map <- ol_map(
    center=c(-100,30),
    zoom=3,
) + 
    public_OSM_basemap()+
    ol_geom_icon(
        some.r.servers,
        r.icon,
        name="R Servers",
        icon.size.scalar='autoscale',
        src.img.width=r.icon.width,
        toggle.control=TRUE,
        tooltip=r.server.names
) 
## Not run: save as HTML and open in browser
# ol_map2HTML(r.map,'R-servers.html')
# browseURL("R-servers.html")
}
\seealso{
\code{\link{ol_map}}, 
\code{\link{ol_geom_polygon}}, 
\code{\link{ol_geom_circle}},
\code{\link{ol_geom_line}}, 
\code{\link{ol_geom_point}}
}
