% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_line.R
\name{ol_geom_line}
\alias{ol_geom_line}
\title{OpenLayers Line Layer}
\usage{
ol_geom_line(line.obj, mapping = ol_aes(), name = NULL, df = NULL,
  toggle.control = FALSE, lwd = 1, ol.lty = list(),
  color = "#000000", label = NULL, label.params = list(),
  tooltip = NULL, tooltip.params = list())
}
\arguments{
\item{line.obj}{SpatialLinesDataFrame, SpatialLines, list
of lines-like objects, or a two-column matrix of longitude-latitude  
coordinates to be used as ordered line object coordinates.}

\item{mapping}{list created by ol_aes.}

\item{name}{character Layer name.}

\item{df}{data.frame with same number of lines-like objects as \code{line.obj}.
Used for aestheic mapping.  Defaults to line.obj@data if 
\code{class(polygon.obj)==SpatialLinesDataFrame)} and df is
not provided.}

\item{toggle.control}{logical indicating whether this layer will have
a visibility toggle.}

\item{lwd}{numeric line feature width. Used only if no \code{lwd}
aesthetic is provided in \code{mapping}}

\item{ol.lty}{(experimental) numeric vector with length > 1, or 
\code{list} of such vectors. Used only if no \code{ol.lty} aesthetic is
provided in \code{mapping}.  See OpenLayers
\href{https://openlayers.org/en/latest/apidoc/module-ol_style_Stroke-Stroke.html}{ol/style/Stroke Documentation},
'lineDash' property for more information.}

\item{color}{character line color string, or vector of color strings.  Used
only if no \code{color} aesthetic is provided in \code{mapping}}

\item{label}{character vector of line feature labels.}

\item{label.params, tooltip.params}{named lists (e.g., \code{list(property=value)}) of 
label and tooltip position and format parameters.  See \link{ol_geom_polygon} documentation.}

\item{tooltip}{character vector of line feature tooltip popups.}
}
\value{
A list object of class \code{Layer.SpatialLine}.
}
\description{
Function to create a line layer to add to an OpenLayers Map object.
}
\details{
This function creates a list object containing the data required 
to generate an OpenLayers vector layer with features using \code{MultiLineString}.  
See Openlayers 
\href{https://openlayers.org/en/latest/apidoc/module-ol_geom_MultiLineString-MultiLineString.html}{MultiLineString Documentation} for details.
}
\section{Aesthetics}{

\itemize{
\item \code{color}
\item \code{lwd}
\item \code{ol_lty} (experimental; See OpenLayers
\href{https://openlayers.org/en/latest/apidoc/module-ol_style_Stroke-Stroke.html}{ol/style/Stroke Documentation},
'lineDash' property for more information.)
}
}

\examples{
line.matrix1 <- matrix(
    c(
        -80.4,-80.4,
        25.78,25.88
    ),
    ncol=2
)
line.matrix2 <- matrix(
    c(
        -80.25,-80.35,
        25.65,25.65
    ),
    ncol=2
)
line.list <- list(line.matrix1,line.matrix2)
line.df <- data.frame(
    direction=c("vertical","horizontal"),
    no=c(1,2)
)
miami.gray.basemap <- ol_map(
    center=c(-80.385790,25.782618),
    zoom=9
    ) + 
   lightgray() 
line.layer <- ol_geom_line(
    line.list,
    mapping=ol_aes(
        color=no,
        lwd=direction
    ),
    df=line.df,
    name="Miami Lines",
    toggle.control=TRUE,
    tooltip=line.df$direction
)
line.color.scale <- ol_scale_color_continuous(
    name="Number",
    display=TRUE
)
line.width.scale <- ol_scale_lwd_discrete(
    lwd.vector=c(
        horizontal=2,
        vertical=4
    ),
    name="Direction",
    display=TRUE
)
line.map.miami <- miami.gray.basemap + 
    line.layer + 
    line.color.scale +
    line.width.scale
\dontrun{
# Output to file and view
ol_map2HTML(
  line.map.miami,
  'miami_lines.html',
  map.heading="Miami Lines"
)
browseURL("miami_lines.html")
}
}
\seealso{
\code{\link{ol_aes}}, \code{\link{ol_map}}, \code{\link{ol_geom_point}}, 
\code{\link{ol_geom_polygon}}, \code{\link{ol_geom_circle}}
}
