% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_point.R
\name{ol_geom_point}
\alias{ol_geom_point}
\title{OpenLayers Point Layer}
\usage{
ol_geom_point(point.obj, mapping = ol_aes(), name = NULL, df = NULL,
  toggle.control = FALSE, fill = "#00FF00", fill.opacity = 1,
  marker = "pin", size = 0.5, label = NULL, label.params = list(),
  tooltip = NULL, tooltip.params = list())
}
\arguments{
\item{point.obj}{SpatialPointsDataframe, SpatialPoints, or a matrix 
containing columns of point longitudes and latitudes, respectively.}

\item{mapping}{list created by ol_aes. Used for aestheic mapping.}

\item{name}{character Layer name.}

\item{df}{data.frame with same number of rows as \code{point.obj} coordinate
matrix.}

\item{toggle.control}{logical indicating whether this layer will have
a visibility toggle.}

\item{fill}{character color string, or vector of color strings.  Used
only if no \code{fill} aesthetic is provided in \code{mapping}}

\item{fill.opacity}{numeric in [0,1]. Controls circle opacity if
no opacity provided in \code{fill} or \code{fill} aesthetic.}

\item{marker}{character.  The 'pin' marker draws map pointers similar to most web map
applications.  The 'dot' or 'point' markers render as cicular points on the map.
Other marker types are not supported by this method.}

\item{size}{numeric point icon size scalar or vector scalars.  Used
only if no \code{size} aesthetic is provided in \code{mapping}.  A value of
1 translates to an icon width of 40 pixels for "pin" markers, or 20 pixels for
"dot" markers.}

\item{label}{character vector of point feature labels.}

\item{label.params, tooltip.params}{named lists (e.g., \code{list(property=value)}) of 
label and tooltip position and format parameters.  See \link{ol_geom_polygon} documentation.}

\item{tooltip}{character vector of point feature tooltip popups.}
}
\value{
A list object of class \code{Layer.SpatialPoint}.
}
\description{
Function to create a points layer to add to an OpenLayers Map object.
}
\details{
This function stores the data required to generate an OpenLayers 
vector layer with features using \code{Point} 
geometries.  
See OpenLayers \href{https://openlayers.org/en/latest/apidoc/module-ol_geom_Point-Point.html}{Point Documentation}
for details.
}
\section{Aesthetics}{

\itemize{
\item \code{fill}
\item \code{size}
}
}

\examples{
point.matrix <- matrix(
    c(
        -80.885+runif(10),
        25.282618+runif(10)
    ),
    ncol=2
)
point.df <- data.frame(
    pt.type=sample(c("A","B"),10,replace=TRUE),
    pt.value=runif(10)*10
)
miami.map <- ol_map(
    center=c(-80.385790,25.782618),
    zoom=9,
) + 
    nga_basemap("WSM")
miami.points <- ol_geom_point(
    point.matrix,
    df=point.df,
    mapping=ol_aes(fill=pt.type,size=pt.value),
    name="Random Points of Interest",
    marker="pin",
    toggle.control=TRUE,
    tooltip=point.df$pt.type
) 
size.scale <- ol_scale_size_continuous(
    display=TRUE,
    draw.fill='green'
)
fill.scale <- ol_scale_fill_discrete(
    c(B='red',A='green'),
    display=TRUE
)
miami.points.map <- miami.map +
    miami.points +
    size.scale +
    fill.scale

## Not Run: output to file and view
# ol_map2HTML(miami.points.map,'Miami_points.html')
# browseURL('Miami_points.html')
}
\seealso{
\code{\link{ol_aes}}, 
\code{\link{ol_map}}, 
\code{\link{ol_geom_polygon}}, 
\code{\link{ol_geom_circle}},
\code{\link{ol_geom_line}},
\code{\link{ol_geom_icon}}
}
