% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_polygon.R
\name{ol_geom_polygon}
\alias{ol_geom_polygon}
\title{OpenLayers Polygon Layer}
\usage{
ol_geom_polygon(polygon.obj, mapping = ol_aes(), name = NULL,
  df = NULL, toggle.control = FALSE, fill = "#00FF00",
  fill.opacity = 0.5, lwd = 1, ol.lty = list(), color = "#000000",
  label = NULL, label.params = list(), tooltip = NULL,
  tooltip.params = list())
}
\arguments{
\item{polygon.obj}{SpatialPolygonsDataFrame, SpatialPolygons, list
of polygon-like objects, or a two-column matrix of longitude-latitude  
coordinates to be used as ordered polygon vertices.}

\item{mapping}{list created by ol_aes.}

\item{name}{character Layer name.}

\item{df}{data.frame with same number of polygon objects as \code{polygon.obj}.
Used for aestheic mapping.  Defaults to polygon.obj@data if 
\code{class(polygon.obj)==SpatialPolygonsDataFrame)} and df is
not provided.}

\item{toggle.control}{logical indicating whether this layer will have
a visibility toggle.}

\item{fill}{character color string, or vector of color strings.  Used
only if no \code{fill} aesthetic is provided in \code{mapping}}

\item{fill.opacity}{numeric in [0,1]. Controls circle opacity if
no opacity provided in \code{fill} or \code{fill} aesthetic.}

\item{lwd}{numeric polygon border width. Used only if no \code{lwd}
aesthetic is provided in \code{mapping}}

\item{ol.lty}{(experimental) numeric vector with length > 1, or 
\code{list} of such vectors. Used only if no \code{ol.lty} aesthetic is
provided in \code{mapping}.  See OpenLayers
\href{https://openlayers.org/en/latest/apidoc/module-ol_style_Stroke-Stroke.html}{ol/style/Stroke Documentation},
'lineDash' property for more information.}

\item{color}{character border color string, or vector of color strings.  Used
only if no \code{color} aesthetic is provided in \code{mapping}}

\item{label}{character vector of polygon feature labels.}

\item{label.params}{named list (e.g., \code{list(property=value)}) of 
label position and format parameters.  See below.}

\item{tooltip}{character vector polygon feature tooltip popups.}

\item{tooltip.params}{named list (e.g., \code{list(property=value)}) of 
tooltip position and format parameters.  See below.}
}
\value{
A list object of class \code{Layer.SpatialPolygon}.
}
\description{
Function to create a polygon layer to add to an OpenLayers Map object.
}
\details{
This function creates a list object containing the data required 
to generate an OpenLayers vector layer with features using \code{MultiPolygon}.  
See OpenLayers 
\href{https://openlayers.org/en/latest/apidoc/module-ol_geom_MultiPolygon-MultiPolygon.html}{MultiPolygon Documentation} for details.
}
\section{Aesthetics}{

\itemize{
\item \code{fill}
\item \code{color}
\item \code{lwd}
\item \code{ol_lty} (experimental; See OpenLayers
\href{https://openlayers.org/en/latest/apidoc/module-ol_style_Stroke-Stroke.html}{ol/style/Stroke Documentation},
'lineDash' property for more information.)
}
}

\section{Formatting Labels With \code{label.params}}{

The \code{label.params} parameter provide direct access to OpenLayers
feature text styling (see \href{http://geoadmin.github.io/ol3/apidoc/ol.style.Text.html}{OpenLayers Documentation}).
Multiple values for any of these properties is not supported.
The following ol/style/Text properties are supported:
\tabular{ll}{
\code{font} \tab character label font CSS string \cr
\code{offsetX} \tab numeric label x-offset \cr
\code{offsetY} \tab numeric label y-offset \cr
\code{rotation} \tab numeric label rotation  \cr
\code{textAlign} \tab character label text horizontal alighment \cr
\code{textBaseline} \tab  character label text vertical alignment\cr
\code{stroke.color} \tab character text color\cr
\code{fill.color} \tab  character text fill color\cr
}
}

\section{Formatting Tooltips With \code{label.params}}{

The \code{tooltip.params} parameter enable the user to control tooltip
formats.  Unlike the \code{label.params}, not all \code{tooltip.params} are 
embedded in Openlayers javascript objects; some are translated
to corresponding CSS properties.  The 
table below provides a list of supported properties and their descriptions.
OpenLayers \href{https://openlayers.org/en/latest/apidoc/module-ol_Overlay.html}{Overlay Documentation}
provides additional information about Overlay Properties.
\tabular{ll}{
\code{font} \tab character tootltip CSS font \cr
\code{offsetX} \tab numeric OpenLayers Overlay x-offset \cr
\code{offsetY} \tab numeric OpenLayers Overlay y-offset \cr
\code{positioning} \tab character OpenLayers Overlay positioning string \cr
\code{stroke.color} \tab character Tooltip CSS font-color \cr
\code{fill.color} character \tab Tooltip CSS background-color \cr
\code{padding}, \tab character Tooltip CSS padding \cr
\code{border}, \tab character Tooltip CSS border \cr
\code{borderradius} character \tab Tooltip CSS border-radius \cr
}
}

\examples{
polygon.matrix1 <- matrix(
    c(
        -80.385+c(0,0.05,0.05,0,0),
        25.782618+c(0,0,0.05,0.05,0)
    ),
    ncol=2
)
polygon.matrix2 <- matrix(
    c(
        -80.34+c(0,0.05,0.025,0),
        25.73++c(0,0,0.025*sqrt(3),0)
    ),
    ncol=2
)
polygon.list<-list(polygon.matrix1,polygon.matrix2)
polygon.df <- data.frame(shape=c("rectangle","triangle"),no=c(1,2))
miami.OSM.basemap <- ol_map(
    center=c(-80.385790,25.782618),
    zoom=9,
    map.heading="Miami Shapes",
    map.note="Note: Mouseover popup values are 
        independent of shape size &amp; color."
    ) + 
   public_OSM_basemap() 
polygon.layer <- ol_geom_polygon(
    polygon.list,
    mapping=ol_aes(
        fill=shape,
    ),
    df=polygon.df,
    lwd=1,
    name="Miami Polygons",
    toggle.control=TRUE,
    tooltip=polygon.df$no
)
polygon.fill.scale <- ol_scale_fill_discrete(display=TRUE)
polygons.over.miami <- miami.OSM.basemap + 
    polygon.layer + 
    polygon.fill.scale

## Not Run: output to file and view
# ol_map2HTML(polygons.over.miami,'miami_polygons.html')
# browseURL("miami_polygons.html")
}
\seealso{
\code{\link{ol_aes}}, \code{\link{ol_map}}, \code{\link{ol_geom_point}}, 
\code{\link{ol_geom_line}}, \code{\link{ol_geom_circle}}
}
