% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_text.R
\name{ol_geom_text}
\alias{ol_geom_text}
\title{OpenLayers Text Layer}
\usage{
ol_geom_text(point.obj, label, name = NULL, toggle.control = FALSE,
  label.params = list(), tooltip = NULL, tooltip.params = list())
}
\arguments{
\item{point.obj}{SpatialPointsDataframe, SpatialPoints, or a matrix 
containing columns of point longitudes and latitudes, respectively.}

\item{label}{character vector of text labels to put at points.}

\item{name}{character Layer name.}

\item{toggle.control}{logical indicating whether this layer will have
a visibility toggle.}

\item{label.params}{named list (e.g., \code{list(property=value)}) of 
label position and format parameters.  See below.}

\item{tooltip}{character vector of point feature tooltip popups.}

\item{tooltip.params}{named list (e.g., \code{list(property=value)}) of 
tooltip position and format parameters.  See \link{ol_geom_polygon} documentation.}
}
\value{
A list object of class \code{Layer.Text}.
}
\description{
Function to create a Text layer to add to an OpenLayers Map object.
}
\details{
This function stores the data required to generate an OpenLayers 
vector layer with text features using \code{Point} 
geometries.  It does not enable aesthetic mappings to variables.
}
\section{Formatting Labels With \code{label.params}}{

The \code{label.params} parameter provide direct access to OpenLayers
feature text styling (see \href{http://geoadmin.github.io/ol3/apidoc/ol.style.Text.html}{OpenLayers Documentation}).
Multiple values for any of these properties is not supported.
The following ol/style/Text properties are supported:
\tabular{ll}{
\code{font} \tab character label font CSS string \cr
\code{offsetX} \tab numeric label x-offset \cr
\code{offsetY} \tab numeric label y-offset \cr
\code{rotation} \tab numeric label rotation  \cr
\code{textAlign} \tab character label text horizontal alighment \cr
\code{textBaseline} \tab  character label text vertical alignment\cr
\code{stroke_color} \tab character text color\cr
\code{fill_color} \tab  character text fill color\cr
}
}

\examples{
text.pts <- matrix(
    c(
        -101.5, 39.2,
        -101.1, 54,
        -101.1, 21.4
    ),
    byrow=TRUE,
    ncol=2
)
text.labels <- c("USA","Canada","Mexico")
mymap <- ol_map(
    center=c(-100,25),
    zoom=3
) +
    oceanbase() +
    ol_geom_text(
        text.pts,
        text.labels,
        toggle.control=TRUE,
        label.params=list(
            font="16px sans-serif",
            stroke_color="#FF0000",
            fill_color="#FFFFFF00"
        )
    )
\dontrun{
# Write to file and view in browser
ol_map2HTML(mymap, "textmap.html")
browseURL("textmap.html")
}
}
\seealso{
\code{\link{ol_map}}, 
\code{\link{ol_geom_point}}
}
