% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ol_map_base.R
\name{ol_map}
\alias{ol_map}
\title{OpenLayers Map}
\usage{
ol_map(zoom = 10, center = c(-117.1611, 32.7157), width = NULL,
  height = NULL,
  ol.source.url = "http://home.gvs.nga.smil.mil/libs/openlayers/3.16.0/build/ol.js",
  map.heading = NULL, map.note = NULL)
}
\arguments{
\item{zoom}{integer map initial zoom level.}

\item{center}{numeric vector of length 2 containing decimal longitude and
latitude coordinates for initial map center.}

\item{width}{numeric or character CSS value width of map container.}

\item{height}{numeric or character CSS value height of map container.}

\item{ol.source.url}{character string containing the url to the OpenLayers
javascript library.}

\item{map.heading}{character heading to be placed over map in html h1 tag.}

\item{map.note}{character note placed in html paragraph (<p>) tag centered
under map container.}
}
\value{
A list object of class \code{Ol.Map}.
}
\description{
Create an OpenLayers Map Object.
}
\details{
This function creates a new S3 OpenLayers Map object with no layers.
}
\examples{
miami.OSM.basemap <- ol_map(
    center=c(-80.385790,25.782618),
    zoom=9,
    map.heading="Miami Shapes",
    map.note="Note: Mouseover popup values are 
        independent of shape size &amp; color."
    ) + 
   nga_basemap('WSM') 
## Not Run
# ol_map2HTML(miami.OSM.basemap,'miami.html')
# browseURL("miami.html")
}
\seealso{
\code{\link{ol_map2HTML}}, 
\code{\link{ol_map2Strings}}, 
\code{\link{nga_basemap}}, 
\code{\link{user_arcgis_basemap}}
}
