% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{ol_map2HTML}
\alias{ol_map2HTML}
\title{Export OpenLayers Map to file.}
\usage{
ol_map2HTML(ol.map.obj, file.name, file.path = ".",
  page.name = "ROpenLayers Map", image.path = "images",
  self.contained = TRUE, nice.format = FALSE,
  IE.compatability.view = TRUE)
}
\arguments{
\item{ol.map.obj}{Ol.Map object to be exported.}

\item{file.name}{character output HTML file name.  This name will be appended
to the \code{file.path}.}

\item{file.path}{character output HTML file path.  The output HTML file and
supporting directories (e.g., 'images') will be in this path.}

\item{page.name}{character page title to be included in the HTML head section.}

\item{image.path}{character \emph{relative} path from \code{file.path} to 
directory that will contain page images.}

\item{self.contained}{logical.  If \code{TRUE} images will be converted to 
64-bit text and embedded in HTML/JavaScript code.  Note: images will still
be created and saved locally in the \code{image.path} folder.}

\item{nice.format}{logical.  If \code{TRUE}, output file will be formated with
new lines and indentation for human readability.}

\item{IE.compatability.view}{logical.  If \code{TRUE}, the statement\cr
\code{<meta http-equiv="X-UA-Compatible" content="IE=edge"/>}\cr
to the HTML document head.  This statement is required for some browsers to 
render the map.}
}
\description{
Writes Ol.Map object to HTML file.
}
\details{
Ol.Map object is written to HTML file with embedded javascript.  The file
will source the OpenLayers javascript library identified by the Ol.Map 
object and call any REST APIs required for each layer.  The intent is to
produce an output file with supporting images, as required, that can
be placed directly into a directory hosted by a minimal http server.
}
\examples{
mymap <- ol_map()
base.layer <- nga_basemap('LightGray')
mymap <- mymap + base.layer
## The following writes HTML and needed images (not run)
# ol_map2HTML(mymap,"SanDiego.html", nice.format=TRUE)
## Open in browser (not run)
# browseURL("SanDiego.html")
}
\seealso{
\code{\link{ol_map}}, 
\code{\link{ol_map2Strings}},
}
