% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ol_scale_color_discrete}
\alias{ol_scale_color_discrete}
\title{Line Color Scale (Discrete)}
\usage{
ol_scale_color_discrete(color.vector = NULL, name = NULL,
  na.col.val = "#FFFFFF00", ordered.values = NULL, opacity = 1,
  preserve.opacity = FALSE, draw.lty = "solid", draw.lwd = 3,
  display = FALSE)
}
\arguments{
\item{color.vector}{character named vector of the form 
\code{c(value=color)}.  If \code{NULL}, a default color mapping
is assigned.}

\item{name}{character the scale name.}

\item{na.col.val}{character the color assigned to unrecognized or
NA values.}

\item{ordered.values}{character, numeric, or factor vector containing the 
ordered unique discrete variable values.  This input is used to determine
the order of the values appearing in the legend.  If not supplied, the 
order is taken from \code{names(color.vector)}.}

\item{opacity}{numeric in [0,1]. The color opacity, if not specified in
the \code{color.vector} colors.}

\item{preserve.opacity}{logical indicating whether to draw the legend
with the same opacity as the feature fills on the map.}

\item{draw.lty}{character indicating the line type for the legend
only.  This will be passed to an R plot command.  See \code{\link{par}}.}

\item{draw.lwd}{numeric width of lines only used in drawing the legend.}

\item{display}{logical indicating whether to draw the scale for output
in the HTML.  If \code{TRUE}, a bitmap will be created and sourced in 
the HTML in an \code{img} tag.  If \code{FALSE}, the \code{draw.*} 
inputs are ignored.}
}
\value{
list of class \code{Scale.Color.Discrete}.
}
\description{
Specify a discrete line color mapping scale.
}
\details{
This method maps OpenLayers feature line colors to discrete 
variable values.  This scale can be added to an Ol.Map S3 object
only if the Ol.Map object has a layer with a "color" mapping to a
numeric, character, or factor type variable.  If no such layer 
exists, attempts to add this type of scale will result in a warning.
The \code{color.vector} input enables the user to specify the exact 
mapping, assigning colors to specific variable values.
}
\examples{
line.matrix1 <- matrix(
    c(
        -80.4,-80.4,
        25.78,25.88
    ),
    ncol=2
)
line.matrix2 <- matrix(
    c(
        -80.25,-80.35,
        25.65,25.65
    ),
    ncol=2
)
line.list <- list(line.matrix1,line.matrix2)
line.df <- data.frame(
    direction=c("vertical","horizontal"),
    no=runif(2)
)
miami.gray.basemap <- ol_map(
    center=c(-80.385790,25.782618),
    zoom=9,
    map.heading="Miami Lines"
    ) + 
   nga_basemap("LightGray") 
line.layer <- ol_geom_line(
    line.list,
    mapping=ol_aes(
        color=direction # discrete mapping
    ),
    df=line.df,
    name="Miami Lines",
    toggle.control=TRUE,
    lwd=5
)
line.color.scale <- ol_scale_color_discrete(
    color.vector=c(
        vertical = 'red',
        horizontal = 'blue'
    ),
    name="Direction",
    display=TRUE
)
line.map.miami <- miami.gray.basemap + 
    line.layer + 
    line.color.scale
## Not Run: output to file and view
# ol_map2HTML(line.map.miami,'miami_lines.html')
# browseURL("miami_lines.html")
}
\seealso{
\code{\link{ol_map}},
\code{\link{ol_geom_polygon}},
\code{\link{ol_geom_line}}
}
