% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ol_map_base.R
\name{public_arcgis_basemap}
\alias{public_arcgis_basemap}
\title{Public ArcGIS Basemap Layer}
\usage{
public_arcgis_basemap(basemap.identifier = "DeLorme", name = NULL,
  toggle.control = FALSE)
}
\arguments{
\item{basemap.identifier}{character indicating which Public ArcGIS mapserver to use.
See 'Available Base Maps'.}

\item{name}{character layer name.}

\item{toggle.control}{logical.  If \code{TRUE}, a checkbox will appear on the
map allowing the viewer to toggle its visibility in the browser.}
}
\value{
A \code{Layer.ArcGIS} S3 object.
}
\description{
Create a basemap layer linking to an Public ArcGIS mapserver.
}
\details{
Creates and returns an OpenLayers ArcGIS Tile layer that sources a
map server hosted at \url{http://server.arcgisonline.com}.  If the 
\code{basemap.identifier} parameter is unrecognized the function will
default to the \href{http://server.arcgisonline.com/ArcGIS/rest/services/Specialty/DeLorme_World_Base_Map/MapServer}{DeLorme} map server.
}
\section{Available Base Maps}{

The following basemap.identifiers are currently supported by this method.
\tabular{ll}{
"LightGray" \tab \href{http://server.arcgisonline.com/arcgis/rest/services/Canvas/World_Light_Gray_Base/MapServer}{World Light Gray Base}\cr
"USAPOP2010" \tab \href{http://server.arcgisonline.com/arcgis/rest/services/Demographics/USA_2000-2010_Population_Change/MapServer}{USA Population Change 2000-2010}\cr
"Hillshade" \tab \href{http://server.arcgisonline.com/arcgis/rest/services/Elevation/World_Hillshade/MapServer}{World Hillshade}\cr
"OceanBase" \tab \href{http://server.arcgisonline.com/arcgis/rest/services/Ocean/World_Ocean_Base/MapServer}{World Ocean Base}\cr
"WorldBoundaries" \tab \href{http://server.arcgisonline.com/arcgis/rest/services/Reference/World_Boundaries_and_Places/MapServer}{World Boundaries and Places}\cr
"WorldRefOverlay" \tab \href{http://server.arcgisonline.com/arcgis/rest/services/Reference/World_Reference_Overlay/MapServer}{World Reference Overlay}\cr
"WorldTrans" \tab \href{http://server.arcgisonline.com/arcgis/rest/services/Reference/World_Transportation/MapServer}{World Transportation}\cr
"WorldNav" \tab \href{http://server.arcgisonline.com/arcgis/rest/services/Specialty/World_Navigation_Charts/MapServer}{World Navigation Charts}\cr
"Imagery" \tab \href{https://server.arcgisonline.com/arcgis/rest/services/World_Imagery/MapServer}{World Imagery} \cr
"DeLorme" \tab \href{http://server.arcgisonline.com/ArcGIS/rest/services/Specialty/DeLorme_World_Base_Map/MapServer}{World Imagery}\cr
}
}

\examples{
mymap <- ol_map()
base.layer <- public_arcgis_basemap('LightGray')
mymap <- mymap + base.layer
## Not run
# ol_map2HTML(mymap,"SanDiego.html")
# browseURL("SanDiego.html")
}
\seealso{
\code{\link{ol_map}}, 
\code{\link{+.Ol.Map}}, 
\code{\link{nga_basemap}}, 
\code{\link{public_OSM_basemap}}, 
\code{\link{user_arcgis_basemap}}
}
