% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ol_map_base.R
\name{ol_map}
\alias{ol_map}
\title{OpenLayers Map}
\usage{
ol_map(zoom = 10, center = c(-117.1611, 32.7157))
}
\arguments{
\item{zoom}{integer map initial zoom level.}

\item{center}{numeric vector of length 2 containing decimal longitude and
latitude coordinates for initial map center.}
}
\value{
A list object of class \code{Ol.Map}.
}
\description{
Create an OpenLayers Map Object.
}
\details{
This function creates a new S3 OpenLayers Map object with no layers.
If \code{ol.source.url} is \code{NULL} and \code{nga.olsource} is
\code{FALSE}, OpenLayers Javascript source will be embedded directly
into the HTML when \code{\link{ol_map2HTML}} or
\code{\link{ol_map2Strings}} is called.
Otherwise, the output HTML/Javascript with source the OpenLayers library
according to the value of \code{ol.source.url}, or the NGA hosted
OpenLayers library if \code{nga.olsource} is \code{TRUE}.
}
\examples{
miami.OSM.basemap <- ol_map(
    center=c(-80.385790,25.782618),
    zoom=9
    ) +
   streetmap()
\dontrun{
ol_map2HTML(
  miami.OSM.basemap,
  'miami.html',
  map.heading="Miami, FL"
)
browseURL("miami.html")
}
}
\seealso{
\code{\link{ol_map2HTML}},
\code{\link{ol_map2Strings}},
\code{\link{nga_basemap}},
\code{\link{user_arcgis_basemap}}
}
